backend/requirements.txt
fastapi[standard]
uvicorn[standard]
google-genai
pydantic
python-dotenv
playwright==1.58.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0
pytest
httpx

